% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{adj2stack}
\alias{adj2stack}
\title{Calculation of network metrics}
\usage{
adj2stack(adjmov, grph = T)
}
\arguments{
\item{adjmov}{Adjacency matrix, need to be an object of class adjmov}

\item{grph}{Whether node level metrics are to be plotted}
}
\value{
A raster stack object
}
\description{
Transform an adjancency matrix to a series of network metrics at the node-level (weight, degree, betweenness, transitivity, eccenctricity) and graph level (diameter, transitivity, density, and modularity)
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
stck<-adj2stack(traj2adj(traj1, res=100), grph=T)
}
\keyword{traj2adj}
