% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{clust_stack}
\alias{clust_stack}
\title{Back-association of population-level clustering to individual clusters}
\usage{
clust_stack(grid, pop_clust, ind_clust, table)
}
\arguments{
\item{grid}{The output of the loop function}

\item{pop_clust}{The output of the pop_clust function}

\item{ind_clust}{The output of the ind_clust function}

\item{table}{The output of table_cluster}
}
\value{
A list of raster stack object.
}
\description{
Generate individual-level rasters of population-level clustering. For each individual, the fuction generate a raster stack containing a raster of the most likely cluster, and several rasters giving the probability of observing each cluster.
}
\examples{
data(albatross)
grid<-loop(albatross, 35000)
table_grid<-table_cluster(albatross, grid)
ls_ind<-ind_clust(table_grid, max.n.clust=8)
pop<-pop_clust(albatross, ls_ind)
clust_stack<-clust_stack(grid, pop, ls_ind, table_grid)
plot(clust_stack[[1]])
}
\keyword{adj2stack}
\keyword{ind_clust}
\keyword{loop}
\keyword{pop_clust}
\keyword{table_cluster}
\keyword{traj2adj}
