% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{graphmet}
\alias{graphmet}
\title{Summarize graph-level metrics}
\usage{
graphmet(grid)
}
\arguments{
\item{grid}{An object generated by the function adj2stack}
}
\value{
A vector
}
\description{
Summarize graph-level metrics from an object generated by adj2stack
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
stck<-adj2stack(traj2adj(traj1, res=quant(traj1)), grph=T)
graphmet(stck)
}
\keyword{adj2stack}
