% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{ind_clust}
\alias{ind_clust}
\title{Individual-level clustering of movement metrics}
\usage{
ind_clust(table, max.n.clust = 8, modelname = "EEV",
  vars = c("Weight", "Degree", "Betweenness", "Speed", "DotP"))
}
\arguments{
\item{table}{An output from the table_cluster function}

\item{max.n.clust}{The maximum number of clusters to test, see the documentation for mclust for more information. Default = 8.}

\item{modelname}{The model structure of the clustering, see the documentation for mclust for more information. Default is equal mean and variance for each clusters (EEV).}

\item{vars}{The variable to be included. Default = c("Weight", "Degree", "Betweenness", "Speed", "DotP")}
}
\value{
A list object with each element representing an individual.
}
\description{
Perform individual-level clustering (first step) of movement metrics. This function uses the output of table_cluster and perform a mixture-model. Users can select which variables will be used and the maximum number of clusters. See also mclust
}
\examples{
data(albatross)
grid<-loop(albatross, 35000)
table_grid<-table_cluster(albatross, grid)
ls_ind<-ind_clust(table_grid, max.n.clust=8)
table(unlist(lapply(ls_ind, function(x) x$G)))
}
\keyword{adj2stack}
\keyword{loop}
\keyword{pop_clust}
\keyword{table_cluster}
\keyword{traj2adj}
