% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{interpolation}
\alias{interpolation}
\title{Interpolation based on movement steps for all individuals}
\usage{
interpolation(traj, ls, wei = mean, deg = mean, bet = max,
  spe = mean, dt = dot)
}
\arguments{
\item{traj}{An object produce by the function adj2stack}

\item{ls}{An object produced by the loop}

\item{wei}{Whether mean or max should be used for weight (default = mean)}

\item{deg}{Whether mean or max should be used for degree (default = mean)}

\item{bet}{Whether mean or max should be used for betweeness (default = max)}

\item{spe}{Whether mean or max should be used for speed (default = mean)}

\item{dt}{Whether mean, max, or dot produc should be used for turning angle (default = dot)}
}
\value{
A list object containing a raster stack object for each individual
}
\description{
Use movement steps to interpolate raster. User can extract the mean or max when multiple steps overlap in a single pixel. Function need to be applied followign the loop function.  This process is very slow
}
\examples{
data(puechabonsp)
locs <- puechabonsp$relocs
xy <- coordinates(locs)
df <- as.data.frame(locs)
da <- as.character(df$Date)
da <- as.POSIXct(strptime(as.character(df$Date),"\%y\%m\%d", tz="Europe/Paris"))
litr <- as.ltraj(xy, da, id = id)
out1<-loop(litr)
out2<-interpolation(litr, out1)
}
\keyword{adj2stack}
\keyword{loop}
\keyword{traj2adj}
