% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{loop}
\alias{loop}
\title{Looping over all individuals}
\usage{
loop(traj, res = 100)
}
\arguments{
\item{traj}{An object produce by the function adehabitatLT with multiple individuals}

\item{res}{Grid size, will be apply to all individuals}
}
\value{
A list object containing a raster stack object for each individual
}
\description{
Extract the adjancency matrix and calculate network metrics for all individuals in a trajectory object. Also calculate mean speed, mean direction, and dot product of turning angles
}
\examples{
data(puechabonsp)
locs <- puechabonsp$relocs
xy <- coordinates(locs)
df <- as.data.frame(locs)
da <- as.character(df$Date)
da <- as.POSIXct(strptime(as.character(df$Date),"\%y\%m\%d", tz="Europe/Paris"))
litr <- as.ltraj(xy, da, id = id)
out1<-loop(litr)
}
\keyword{adj2stack}
\keyword{traj2adj}
