% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{mosaic_network}
\alias{mosaic_network}
\title{Mosaic individuals together for a given variable}
\usage{
mosaic_network(ls, index = 2, sc = T, fun = mean)
}
\arguments{
\item{ls}{An object produced by the loop or interpolate functions}

\item{index}{Index indicating which layer to take in the stack}

\item{sc}{Whether to scale all individual rasters (default = TRUE)}

\item{fun}{Whether mean or max should be used as the mosaic function (default = mean)}
}
\value{
A list object containing a raster stack object for each individual
}
\description{
Use output of loop or interpolation and combine all individuals (mosaic) together using the mean or max values.
}
\examples{
data(puechabonsp)
locs <- puechabonsp$relocs
xy <- coordinates(locs)
df <- as.data.frame(locs)
da <- as.character(df$Date)
da <- as.POSIXct(strptime(as.character(df$Date),"\%y\%m\%d", tz="Europe/Paris"))
litr <- as.ltraj(xy, da, id = id)
out1<-loop(litr)
mean_weight<-mosaic_network(out1, index=2, sc=T, fun=mean) #Perform mean weight (not-interpolated)
plot(mean_weight)
}
\keyword{adj2stack}
\keyword{loop}
\keyword{traj2adj}
