% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{pop_clust}
\alias{pop_clust}
\title{Population-level clustering of movement metrics}
\usage{
pop_clust(traj, ls, max.n.clust = 8)
}
\arguments{
\item{traj}{The trajectory object}

\item{max.n.clust}{The maximum number of clusters to test, see the documentation for mclust for more information. Default = 8.}

\item{ls_ind}{Individual-level clustering object, the output of ind_clust.}
}
\value{
A list object with each element representing an individual.
}
\description{
Combine individual-level clustering of movement metrics into a population-level clustering (second step). Users can  the maximum number of clusters. See also mclust
}
\examples{
data(albatross)
grid<-loop(albatross, 35000)
table_grid<-table_cluster(albatross, grid)
ls_ind<-ind_clust(table_grid, max.n.clust=8)
pop<-pop_clust(albatross, ls_ind)
pop[[1]]$parameters$mean
pop[[1]]$parameters$pro
}
\keyword{adj2stack}
\keyword{ind_clust}
\keyword{loop}
\keyword{table_cluster}
\keyword{traj2adj}
