% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{pop_overl}
\alias{pop_overl}
\title{Population-level multi-use map}
\usage{
pop_overl(clust_stack)
}
\arguments{
\item{clust_stack}{The output of clust_stack}
}
\value{
A stack object with each raster showing use of each cluster
}
\description{
Produce a single map (raster) indicating all types of use found in a cluster.
}
\examples{
data(albatross)
grid<-loop(albatross, 35000)
table_grid<-table_cluster(albatross, grid)
ls_ind<-ind_clust(table_grid, max.n.clust=8)
pop<-pop_clust(albatross, ls_ind)
clust_stack<-clust_stack(grid, pop, ls_ind, table_grid)
pop_overl<-pop_overl(clust_stack)
table(values(pop_overl))
plot(pop_overl)
}
\keyword{adj2stack}
\keyword{ind_clust}
\keyword{loop}
\keyword{table_cluster}
\keyword{traj2adj}
