% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{pop_stack}
\alias{pop_stack}
\title{Population-level single-use maps of each cluster}
\usage{
pop_stack(clust_stack)
}
\arguments{
\item{clust_stack}{The output of clust_stack}
}
\value{
A stack object with each raster showing use of each cluster
}
\description{
Produce maps (raster) indicating if at least one individual is using a pixel as a given cluster
}
\examples{
data(albatross)
grid<-loop(albatross, 35000)
table_grid<-table_cluster(albatross, grid)
ls_ind<-ind_clust(table_grid, max.n.clust=8)
pop<-pop_clust(albatross, ls_ind)
clust_stack<-clust_stack(grid, pop, ls_ind, table_grid)
pop_stack<-pop_stack(clust_stack)
plot(pop_stack)
}
\keyword{adj2stack}
\keyword{ind_clust}
\keyword{loop}
\keyword{table_cluster}
\keyword{traj2adj}
