% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{quant}
\alias{quant}
\title{Sample quantile of distance for ltraj object}
\usage{
quant(x, p = 0.5)
}
\arguments{
\item{x}{A ltraj object}

\item{p}{Probability, default=0.5 (median)}
}
\value{
A vector of length p
}
\description{
Wrapper function that extract the sample quantile of distance
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
stck<-adj2stack(traj2adj(traj1, res=quant(traj1)), grph=T)
}
\keyword{ltraj}
