% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{sim_mov}
\alias{sim_mov}
\title{Simulation of patch-based movement trajectory}
\usage{
sim_mov(type = c("2states", "OU"), npatches = 5, ratio = 5,
  nswitch = 150, ncore = 200, spacecore = 200,
  seq_visit = sample(1:npatches, nswitch, replace = T),
  stepDist = "gamma", angleDist = "vm", stepPar = c(0.5, 3, 1, 5),
  anglePar = c(pi, 0, 0.5, 2), s = diag(40, 2), grph = F)
}
\arguments{
\item{type}{whether movement within patches should be based on a 2states process (from package moveHMM) or a Bivariate Ornstein-Uhlenbeck process (OU) (from package adehabitatLT)}

\item{npatches}{Number of patches, default=5}

\item{ratio}{Ratio (in percent) of locations associated to interpatch movement, default=5}

\item{nswitch}{Number of switch/depart from patches, default=150}

\item{ncore}{Number of locations within a patch per visit, default=200}

\item{spacecore}{Minimum distance between center of patches, default=200}

\item{seq_visit}{Specify the sequence of visit among patches, default is random sequence}

\item{stepDist}{Distribution for step length if 2states specified in type, see simData of moveHMM package}

\item{angleDist}{Distribution for turn angle if 2states specified in type, see simData of moveHMM package}

\item{stepPar}{Parameters for step length distribution if 2states specified in type, see simData of moveHMM package}

\item{anglePar}{Parameters for turn angle distribution if 2states specified in type, see simData of moveHMM package}

\item{s}{Parameters for the OU process, see simm.mou of adehabitatLT package}

\item{grph}{Whether a graph of the trajectory should be produced, default=F}
}
\value{
A ltraj (adehabitatLT) object
}
\description{
Simulate a movement trajectory with user defined number of patches and interpatch movement
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
traj2<-sim_mov(type="2states", npatches=2, grph=T)
}
\keyword{adj2stack}
\keyword{traj2adj}
