% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{table_cluster}
\alias{table_cluster}
\title{Convert a list of adj2stack object to a data.frame for clustering}
\usage{
table_cluster(traj, grid)
}
\arguments{
\item{traj}{The trajectory used in loop (a traj object)}

\item{grid}{The output of the loop function}
}
\value{
A data.frame object.
}
\description{
Convert output of loop function to a data.frame.
}
\examples{
data(albatross)
grid<-loop(albatross, 35000)
table_grid<-table_cluster(albatross, grid)
head(table_grid)
}
\keyword{adj2stack}
\keyword{loop}
\keyword{traj2adj}
