% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{traj2adj}
\alias{traj2adj}
\title{Generation of adjacency matrix from movement data}
\usage{
traj2adj(mov, res = 100, grid = NULL)
}
\arguments{
\item{mov}{Movement trajectory, need to be a ltraj object}

\item{res}{Grid size (based on coordinate system of movement trajectory)}

\item{grid}{User specified grid (a raster), needs to have a larger extent than the movement trajectory}
}
\value{
A list of objects containing the adjacency matrix, the grid use, and patch/corridor identification (only useful if sim_mov was used)
}
\description{
Transform an ltraj object to an adjacency matrix using a user-specified grid size
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
adj<-traj2adj(traj1, res=100)
}
\keyword{adj2stack}
