% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcts packages.R
\name{val}
\alias{val}
\title{Extract occupied cells in a raster object}
\usage{
val(grid, id)
}
\arguments{
\item{grid}{An object generated by the function adj2stack}

\item{id}{Metric to be used (2=Weight, 3=Degree, 4=Betweenness, 5=Transitivity, 6=Eccentricity)}
}
\value{
A vector
}
\description{
Extract only occupied cells in a raster object,
}
\examples{
traj1<-sim_mov(type="OU", npatches=3, grph=T)
stck<-adj2stack(traj2adj(traj1, res=quant(traj1)), grph=T)
mean(val(stck, 2))
}
\keyword{adj2stack}
